+++ src/app/dashboard/dashboard-routing.module.ts
@@ -1,0 +1,18 @@
+import {NgModule} from '@angular/core';
+import {RouterModule, Routes} from '@angular/router';
+
+import {DashboardPage} from './dashboard.page';
+
+const routes: Routes = [
+  {
+    path: '',
+    component: DashboardPage
+  }
+];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule],
+})
+export class DashboardPageRoutingModule {
+}