+++ src/enums/mills/millActions.ts
@@ -1,0 +1,6 @@
+export enum MILL_ACTION {
+
+  INFORMATION = 'INFORMATION',
+  EDIT = 'EDIT',
+  DELETE = 'DELETE',
+}